refactor: move API contracts from shared/ to server/
API contracts (InvoicePatchSchema, ExpectedInvoicePatchSchema, formatZodError)
are server-only concerns and don't belong in shared code. This keeps shared/
focused on types and validators that are truly needed by both client and server.

- Moved Zod schemas to server/contracts/schemas.ts
- Moved formatZodError to server/contracts/utils.ts
- Deleted shared/src/contracts/ entirely
- Updated imports in SvelteKit API endpoints
