+++ client/src/routes/api/expected-invoices/[id]/+server.ts
@@ -5,7 +5,7 @@
 import { json } from '@sveltejs/kit';
 import type { RequestHandler } from './$types';
 import { ExpectedInvoiceRepository } from '@server/database/repositories/expected-invoice.js';
-import { ExpectedInvoicePatchSchema, formatZodError } from '@shared/contracts';
+import { ExpectedInvoicePatchSchema, formatZodError } from '@server/contracts';
 
 /**
  * GET /api/expected-invoices/:id
