+++ shared/src/index.ts
@@ -1,12 +1,9 @@
 /**
- * Shared module - Types, contracts, and validators shared between client and server.
+ * Shared module - Types and validators shared between client and server.
  *
  * @example
  * // Import types
  * import type { Invoice, Currency } from '@shared/types';
- *
- * // Import contracts
- * import { InvoicePatchSchema, formatZodError } from '@shared/contracts';
  *
  * // Import validators
  * import { validateCUIT, normalizeCUIT } from '@shared/validators';
@@ -14,5 +11,4 @@
 
 // Re-export everything for convenience
 export * from "./types/index.js";
-export * from "./contracts/index.js";
 export * from "./validators/index.js";
